+++ NEWS.md
@@ -1,3 +1,8 @@
+# repo.data 0.1.5
+
+* Package will not generate errors if no internet connection is available.
+* Fix error with `update_packages()` with local packages.
+
 # repo.data 0.1.0
 
 * Manually modify pkgdown index.